requirements.txt
uvicorn
fastapi
google-generativeai

# test
pytest